# docs/openapi.yaml
# ============================================================================
# Documentacion de la API - OpenAPI 3.1
#
# TODA la documentacion se escribe en ESTE archivo, nunca en los routers.
# Los routers quedan solo con logica de ruteo.
#
# Se ve en:  http://localhost:3000/api/docs                (local)
#            https://api-instituto.onrender.com/api/docs   (produccion)
#
# Al agregar un endpoint: sumar su tag en `tags` y su ruta en `paths`.
# Abajo de `paths` hay un ejemplo completo comentado para copiar y pegar.
# ============================================================================

openapi: 3.1.0

info:
  title: API Instituto
  version: 1.0.0
  description: |
    API REST del Instituto.

    ### Forma de las respuestas

    **En exito el body es el recurso pelado**, sin ningun sobre: un objeto en
    las respuestas de un recurso y un array en los listados. Un listado vacio
    devuelve `200 []`, nunca un 404. Un `DELETE` devuelve `204` sin body.

    **En error** el body es un objeto `{ success: false, message, details? }`,
    igual para todos los endpoints porque sale del manejador centralizado
    (`src/middlewares/error.middleware.js`).

    Son formas distintas a proposito: para saber si algo salio bien mirá el
    **status HTTP**, no una propiedad del body.

    Las respuestas de error estan definidas una sola vez en
    `components/responses`: no las redefinas, usalas con `$ref`.

servers:
  - url: https://api-instituto.onrender.com/api
    description: Produccion
  - url: http://localhost:3000/api
    description: Desarrollo local

# Un tag por recurso. Agrupa los endpoints en la UI y define el orden en que
# aparecen.
tags:
  - name: Sedes
    description: Sedes donde se dictan las carreras

# Las rutas se escriben SIN el prefijo `/api`: ya esta incluido en `servers`.
# Es decir, el endpoint GET /api/sedes se documenta aca como `/sedes`.
#
# `/sedes` es la referencia a copiar para los recursos que falten.
paths:

  /sedes:
    post:
      tags: [Sedes]
      summary: Crea una sede
      description: |
        No puede haber dos sedes con el mismo `nombre` en la misma `ciudad`.
        Los textos se recortan (trim) antes de guardarse.

        Hoy es publico. Cuando exista el modulo de auth va a requerir un token
        con rol ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SedeInput" }
      responses:
        "201":
          description: Sede creada
          headers:
            Location:
              description: URL de la sede recien creada
              schema: { type: string }
              example: /api/sedes/clx1a2b3c4d5e6f7g8h9i0j1
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sede" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Ya existe una sede con ese nombre en esa ciudad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }

  /sedes/{id}:
    get:
      tags: [Sedes]
      summary: Obtiene una sede por id
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la sede
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "200":
          description: La sede pedida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sede" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

components:

  securitySchemes:
    # Es lo que valida `authenticate` en src/middlewares/auth.middleware.js.
    # Para probar endpoints protegidos desde la UI: boton "Authorize" y pegar
    # el token SIN el prefijo "Bearer " (la UI lo agrega sola).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    # Un schema por entidad, a medida que se implementan. Convencion: `X` es
    # lo que devuelve la API y `XInput` lo que se recibe en el body.

    Sede:
      type: object
      description: Una sede del instituto.
      required: [id, nombre, ciudad, provincia, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: cuid generado por la base.
          examples: ["clx1a2b3c4d5e6f7g8h9i0j1"]
        nombre:
          type: string
          examples: ["Sede Centro"]
        ciudad:
          type: string
          examples: ["Cordoba"]
        provincia:
          type: string
          examples: ["Cordoba"]
        direccion:
          type: [string, "null"]
          examples: ["Av. Colon 1234"]
        telefono:
          type: [string, "null"]
          examples: ["+54 351 4567890"]
        email:
          type: [string, "null"]
          format: email
          examples: ["centro@instituto.edu.ar"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SedeInput:
      type: object
      description: >-
        Body para crear una sede. No acepta campos que no esten en esta lista:
        si mandas uno de mas, la respuesta es un 400.
      required: [nombre, ciudad, provincia]
      additionalProperties: false
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 120
          examples: ["Sede Centro"]
        ciudad:
          type: string
          minLength: 1
          maxLength: 80
          examples: ["Cordoba"]
        provincia:
          type: string
          minLength: 1
          maxLength: 80
          examples: ["Cordoba"]
        direccion:
          type: string
          minLength: 1
          maxLength: 200
          examples: ["Av. Colon 1234"]
        telefono:
          type: string
          minLength: 1
          maxLength: 30
          examples: ["+54 351 4567890"]
        email:
          type: string
          format: email
          maxLength: 120
          examples: ["centro@instituto.edu.ar"]

    Error:
      type: object
      description: Forma de cualquier respuesta de error de la API.
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          description: Descripcion del error, apta para mostrar al usuario.
          examples: ["El registro solicitado no existe"]
        details:
          description: Informacion extra. Solo aparece en algunos errores.
        stack:
          type: string
          description: >-
            Traza del error. SOLO se incluye cuando NODE_ENV=development;
            en produccion nunca se expone.

    ValidationIssue:
      type: object
      description: Un campo que no paso la validacion de Zod.
      required: [source, field, message]
      properties:
        source:
          type: string
          enum: [body, params, query]
          description: En que parte del request estaba el campo.
        field:
          type: string
          examples: ["email"]
        message:
          type: string
          examples: ["Invalid email address"]

    ValidationError:
      type: object
      description: >-
        Error 400 devuelto por el middleware `validate()`. A diferencia del
        resto, `details` trae la lista de campos que fallaron.
      required: [success, message, details]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          examples: ["Datos de entrada invalidos"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"

  responses:

    # Respuestas comunes a todos los endpoints. Se referencian asi:
    #   "404": { $ref: "#/components/responses/NotFound" }

    BadRequest:
      description: Los datos enviados no pasaron la validacion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"

    Unauthorized:
      description: Falta el token, es invalido, expiro o la cuenta esta desactivada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Forbidden:
      description: Autenticado, pero el rol no alcanza para esta accion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: El recurso no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Conflict:
      description: >-
        Choca con un registro existente: campo unico duplicado (Prisma P2002) o
        una relacion que impide la operacion (P2003).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    InternalError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"